move syntax checker into component
